Refactor project structure and extract reusable hooks
